Move all opportunities code to modules
