--- src/apps/investments/client/opportunities/Details/state.js
@@ -1,13 +1,0 @@
-export const TASK_GET_OPPORTUNITY_DETAILS = 'TASK_GET_OPPORTUNITY_DETAILS'
-export const TASK_GET_OPPORTUNITY_DETAILS_METADATA =
-  'TASK_GET_OPPORTUNITY_DETAILS_METADATA'
-export const TASK_GET_OPPORTUNITY_REQUIREMENTS_METADATA =
-  'TASK_GET_OPPORTUNITY_REQUIREMENTS_METADATA'
-
-export const TASK_SAVE_OPPORTUNITY_DETAILS = 'TASK_SAVE_OPPORTUNITY_DETAILS'
-export const TASK_SAVE_OPPORTUNITY_REQUIREMENTS =
-  'TASK_SAVE_OPPORTUNITY_REQUIREMENTS'
-
-export const ID = 'opportunityDetails'
-
-export const state2props = (state) => state[ID]
