+++ src/client/modules/Investments/Opportunities/Opportunity.jsx
@@ -4,14 +4,14 @@ import { connect } from 'react-redux'
 import { Route, useParams } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { DefaultLayout, Main } from '../../../../../client/components'
+import { DefaultLayout, Main } from '../../../components'
 import OpportunityDetails from './OpportunityDetails'
 import OpportunityDetailsHeader from './OpportunityDetailsHeader'
-import TabNav from '../../../../../client/components/TabNav'
-import urls from '../../../../../lib/urls'
-import Task from '../../../../../client/components/Task/index.jsx'
+import TabNav from '../../../components/TabNav'
+import urls from '../../../../lib/urls'
+import Task from '../../../components/Task/index.jsx'
 import { TASK_GET_OPPORTUNITY_DETAILS, ID, state2props } from './state'
-import { INVESTMENT_OPPORTUNITY_DETAILS__LOADED } from '../../../../../client/actions'
+import { INVESTMENT_OPPORTUNITY_DETAILS__LOADED } from '../../../actions'
 import OpportunityInteractions from './OpportunityInteractions'
 
 const StyledMain = styled(Main)`
