+++ src/client/modules/Investments/Opportunities/OpportunityDetailsTable.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import Link from '@govuk-react/link'
 import styled from 'styled-components'
 
-import SummaryTable from '../../../../../client/components/SummaryTable'
-import Tag from '../../../../../client/components/Tag'
+import SummaryTable from '../../../components/SummaryTable'
+import Tag from '../../../components/Tag'
 
 const StyledTag = styled(Tag)`
   float: right;
