+++ src/client/modules/Investments/Opportunities/OpportunityRequirementsTable.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import SummaryTable from '../../../../../client/components/SummaryTable'
+import SummaryTable from '../../../components/SummaryTable'
 
 const OpportunityRequirementsTable = ({ details }) => {
   const {
