+++ src/client/modules/Investments/Opportunities/constants.js
