+++ src/client/modules/Investments/Opportunities/state.js
@@ -2,4 +2,16 @@ export const TASK_CREATE_INVESTMENT_OPPORTUNITY =
   'TASK_CREATE_INVESTMENT_OPPORTUNITY'
 export const TASK_SAVE_OPPORTUNITY_STATUS = 'TASK_SAVE_OPPORTUNITY_STATUS'
 
+export const TASK_GET_OPPORTUNITY_DETAILS = 'TASK_GET_OPPORTUNITY_DETAILS'
+export const TASK_GET_OPPORTUNITY_DETAILS_METADATA =
+  'TASK_GET_OPPORTUNITY_DETAILS_METADATA'
+export const TASK_GET_OPPORTUNITY_REQUIREMENTS_METADATA =
+  'TASK_GET_OPPORTUNITY_REQUIREMENTS_METADATA'
+
+export const TASK_SAVE_OPPORTUNITY_DETAILS = 'TASK_SAVE_OPPORTUNITY_DETAILS'
+export const TASK_SAVE_OPPORTUNITY_REQUIREMENTS =
+  'TASK_SAVE_OPPORTUNITY_REQUIREMENTS'
+
 export const ID = 'opportunityDetails'
+
+export const state2props = (state) => state[ID]
